Toy System Implemented
Modifications to the pet in order to add more depth to the pets individuality with specific "pet types". Represents item compatibility. Toys now effect happiness, current status, and newly added energy according to the mood and compatibility of the pet. Also detects if the pet is too hungry or tired to play.